Example MESA folder, containing modified files.
This MESA folder is provided to serve as a general example of how a 2.5 solar mass model of initial solar metallicity would be evolved through the post-AGB phase while experiencing time-varying accretion from a circumbinary disk. This folder contains a number of modifications made by Kayla Martin for this work, as described in the supplied modifications_summary.txt document. Please note that before use of the contents of this folder,